feat: add sidecar for mounting secrets (#1847)
* add sidecar for mounting secrets

* enable ruff in ci

* feat: implement secrets handling (#1853)

* feature(server): add support for user secrets

The secrets are the ones the user are storing in the secrets storage
service.

The notebook service will request the creation of a k8s secret from the
service that will contain all the secrets to be used in the session.

* feature: add support for user secrets in the API

A new optional field named user_secrets has been added to the server
creation endpoint that will trigger the k8s secret creation upon
successful JupyterServer creation.

This allows to tie the secret to the JupyterServer object and thus
automate its deletion upon jupyter server deletion.

---------

Co-authored-by: Samuel Gaist <samuel.gaist@idiap.ch>