feat: add connected services (#179)
Add the `connected services` module which allows users to connect to third-party service providers. This new feature allows users to connect to externally hosted GitLab instances and to use repositories in sessions.

For example, if the `gitlab.com` provider is configured and the user is connected to it, sessions can start with repositories mounted from `gitlab.com`.