chore(ci): do not start kind on devcontainer startup (#323)
* chore(ci): do not create kind cluster or install amalthea operators

These are not yet used and they are slowing down the startup time of all
features and also in some cases causing the ci to crash and fail when it
tries to start the devcontainer.

* chore(ci): do not set a subnet range for the docker network for kind

If this is specified the devcontainer does not run on my computer. When
this is ommited there is some default range that docker uses.