refactor(tests): move all cluster requiring test in own class
While most of the tests are written in functions, moving those
requiring a cluster into a class allows the use of a class
scoped fixtures for the cluster creation. This has the advantage
of reducing the run time of the tests since the cluster will only
be created for tests that actually requires one so if someone
needs to re-test a particular endpoint that has nothing to do
with kubernetes, they will not incur the useless creation and
destruction time which is not trivial.

The class has been put in utils because it will be used later
for the sessions tests.