fix: accept only lowercase slugs
This also makes it so that we can return uppercase slugs for projects -
because we did not fully lowercase and check for this before. But now we
cannot save or modify any existing project slugs that have uppercase in
them. Whereas if there are any projects with uppercase slugs we can
still list, patch, delete them.