fix: allow project editors to send patches with the current namespace (#483)
With the current Renku 2.0 UI, when a user with the "editor" role tries to update a project, e.g. the description or keywords, they are not allowed to do so because the UI sends a patch containing the current namespace.

This is fixed in this PR, the added test fails prior to the fix.