fix: make work and mount dir optional
We should use the image work dir - so leaving it blank uses the docker
image work dir. The mount dir can be empty because we can default to the
work dir when it is not set. This allows us to not require so many
inputs from users when creating environments.