feat: add support for session secret slots (#529)
Add support for session secret slots which describe files with secret contents to be mounted in sessions. Session secret slots exist at the project level and can be managed by users with at the `WRITE` permission on the project. Session secrets can be used by all users with the `READ` permission on the project.