refactor(shipwright): replace custom helper with lifecycle's creator
The lifecycle repository provides the creator binary which
runs each step of the build lifecycle in a similar fashion as
what the original script did as well as its go counterpart.

This commit thus removes all of that and updates the manifest
to use creator. Python tests have been added to ensure that
the buildrun is successfull.