Change solr documents to carry namespace and path
This is necessary to disambiguate living inside a namespace and being
a namespace. The entity tree is build like this:

- Projects always have a namespace (are never top-level), and they may act
  as a namespace for data connectors
- Groups never have a namespace (are not nested, always top-level),
  they are a namespace for projects and data connectors
- Users never have a namespace (always top-level), they are a
  namespace for projects and data connectors
- DataConnectors *may* have a namespace (are either in a group or
  project, or top-level), they don't act as a namespace for other
  things

In order to resolve namespaces it must be reliably found in the solr
documents, which are flat. It requires a globally unique key. The data
services doesn't carry the internal unique id (ulid or uuid), but the
"path". The path to an entity is always defined and is either
its *slug* (if top-level) or consists of the path of its namespace and
its slug. For resolving namespaces the path can be used and is now
added to each document.

The namespace field is removed. The field `isNamespace` is added with
a static value of `true` to those entities that may act like a
namespace. It is not strictly required but is an extra safety net when
resolving.