chore: Improve logging setup (#873)
Introduces a logging setup to centralize its configuration and removes
changes to the logging config from other places in the code base.

The logging config is called in the bases __init__ file to be as early
as possible applied. For sanic applications, it must also run for
workers and the main process. The configuration sets up logging by
default as follows:

- our code is logged at level INFO
- all third-party code is logged at level WARNING

What is "our" code is determined by the logger name: our loggers
should all be children of the renku_data_services logger. Using the
provided getLogger function makes sure this is the case.

The config can be adjusted using environment variables, as written in
`Config.from_env()`.