Fix for using GitHub as provider for code repos and image registry (#1028)
A GitHub App cannot be used to generate tokens for accessing the image registry at
ghcr.io (see [this discussion](https://github.com/orgs/community/discussions/24636) 
and [this issue](https://github.com/github/roadmap/issues/558). However, it works using
(the older, almost deprecated?) Github OAuth apps.

This is change doesn't introduce another `ProviderKind`, but figures from how the github
provider is given whether it is for accessing images or code repos. A github provider with
a `image_registry_url` present is used only for image access and is expected to correspond
to a GitHub OAuth app. But when a GitHub provider has no `image_registry_url` set, it is 
used for accessing code repositories and is expected to correspond to a GitHub App.

---------
Co-authored-by: Flora Thiebaut <flora.thiebaut@sdsc.ethz.ch>