feat: indicate supported platforms when checking a session image (#1101)
Add a new field `platforms` returned when checking a session image with `GET /api/data/sessions/images`. The field may be returned when the image is accessible and will contain the list of supported platforms. If gathering this information fails, the field will simply not be present in the response.