use database lock to guard updating stale values
Sometimes, github makes it possible to refresh using the same refresh
token. then we would update both in our database, which then results
in one client having the wrong token (because it was overwritten
immediately afterwards). For this to work, the `update_token` hook
cannot be used (easily) as we need more control about the updating
process.

While AsyncOAuth2client uses an asyncio-lock, since it is an instance
variable, it doesn't guard across multiple instance using the same token.