fix: send indentify() call only when necessary (#1115)
Closes #1112.

Send the `identify()` call for metrics only in one of the following conditions:
* The saved identity hash does not exist for the user
* The saved identity hash is outdated

The identity hash is persisted only when metrics are enabled, which means that the users' identity will be backfilled once metrics are enabled.

Also, fix the user sync logic to update users only if the current data is different from the existing data.