feat: describe runtime platform on resource pools (#1116)
⚠️ Breaking API change on `PUT /resource_pools/{resource_pool_id}`: a new `platform` field is required.

Add a `platform` field to resource pools. The `platform` field describes the platform which will run the sessions in the resource pool, e.g `linux/amd64` or `linux/arm64`. The platform information can be checked against the image information to know if the image is compatible with the resource pool.