Remove manifest from str representation
When logging/printing `K8sObject`s, the manifest was included which
can result in very large strings, containing potential sensitive data.
This string would be returned to clients inside an error message (see
referenced issue). The name and namespace should be enough, so the
manifest is now hidden - same as done already in `K8sSecret`.

Fixes: #1127